+++ packages/core/src/blade/common/api/params.ts
@@ -1,14 +1,67 @@
-import {BaseParams} from '../../../common/params';
-import {BooleanInputParams} from '../../../input-binding/boolean/plugin';
-import {ColorInputParams} from '../../../input-binding/color/util';
-import {NumberInputParams} from '../../../input-binding/number/plugin';
-import {Point2dInputParams} from '../../../input-binding/point-2d/plugin';
-import {Point3dInputParams} from '../../../input-binding/point-3d/plugin';
-import {Point4dInputParams} from '../../../input-binding/point-4d/plugin';
-import {StringInputParams} from '../../../input-binding/string/plugin';
-import {BooleanMonitorParams} from '../../../monitor-binding/boolean/plugin';
-import {NumberMonitorParams} from '../../../monitor-binding/number/plugin';
-import {StringMonitorParams} from '../../../monitor-binding/string/plugin';
+import {Formatter} from '../../../common/converter/formatter';
+import {
+	BaseInputParams,
+	BaseMonitorParams,
+	BaseParams,
+	ListParamsOptions,
+	NumberTextInputParams,
+	PickerLayout,
+	PointDimensionParams,
+} from '../../../common/params';
+import {ColorType} from '../../../input-binding/color/model/color-model';
+
+export interface BooleanInputParams extends BaseInputParams {
+	options?: ListParamsOptions<boolean>;
+}
+
+export interface ColorInputParams extends BaseInputParams {
+	color?: {
+		alpha?: boolean;
+		type?: ColorType;
+	};
+	expanded?: boolean;
+	picker?: PickerLayout;
+}
+
+export interface NumberInputParams
+	extends BaseInputParams,
+		NumberTextInputParams {
+	options?: ListParamsOptions<number>;
+}
+
+export interface Point2dYParams extends PointDimensionParams {
+	inverted?: boolean;
+}
+
+export interface Point2dInputParams
+	extends BaseInputParams,
+		PointDimensionParams {
+	expanded?: boolean;
+	picker?: PickerLayout;
+	x?: PointDimensionParams;
+	y?: Point2dYParams;
+}
+
+export interface Point3dInputParams
+	extends BaseInputParams,
+		PointDimensionParams {
+	x?: PointDimensionParams;
+	y?: PointDimensionParams;
+	z?: PointDimensionParams;
+}
+
+export interface Point4dInputParams
+	extends BaseInputParams,
+		PointDimensionParams {
+	x?: PointDimensionParams;
+	y?: PointDimensionParams;
+	z?: PointDimensionParams;
+	w?: PointDimensionParams;
+}
+
+export interface StringInputParams extends BaseInputParams {
+	options?: ListParamsOptions<string>;
+}
 
 type InputParams =
 	| BooleanInputParams
@@ -18,10 +71,37 @@ type InputParams =
 	| Point3dInputParams
 	| Point4dInputParams
 	| StringInputParams;
+
+export interface BooleanMonitorParams extends BaseMonitorParams {
+	/**
+	 * Number of rows for visual height.
+	 */
+	rows?: number;
+}
+
+export interface NumberMonitorParams extends BaseMonitorParams {
+	format?: Formatter<number>;
+	max?: number;
+	min?: number;
+	/**
+	 * Number of rows for visual height.
+	 */
+	rows?: number;
+}
+
+export interface StringMonitorParams extends BaseMonitorParams {
+	multiline?: boolean;
+	/**
+	 * Number of rows for visual height.
+	 */
+	rows?: number;
+}
+
 type MonitorParams =
 	| BooleanMonitorParams
 	| NumberMonitorParams
 	| StringMonitorParams;
+
 export type BindingParams = InputParams | MonitorParams;
 
 export interface ButtonParams extends BaseParams {
