+++ packages/core/src/input-binding/point-3d/plugin.ts
@@ -1,7 +1,7 @@
+import {Point3dInputParams} from '../../blade/common/api/params';
 import {Constraint} from '../../common/constraint/constraint';
 import {parseNumber} from '../../common/converter/number';
 import {parseRecord} from '../../common/micro-parsers';
-import {BaseInputParams, PointDimensionParams} from '../../common/params';
 import {createPointAxis} from '../../common/point-nd/point-axis';
 import {
 	createDimensionConstraint,
@@ -15,14 +15,6 @@ import {PointNdTextController} from '../common/controller/point-nd-text';
 import {InputBindingPlugin} from '../plugin';
 import {point3dFromUnknown, writePoint3d} from './converter/point-3d';
 import {Point3d, Point3dAssembly, Point3dObject} from './model/point-3d';
-
-export interface Point3dInputParams
-	extends BaseInputParams,
-		PointDimensionParams {
-	x?: PointDimensionParams;
-	y?: PointDimensionParams;
-	z?: PointDimensionParams;
-}
 
 function createConstraint(
 	params: Point3dInputParams,
